requirements.txt
# Core Libraries
openai==1.14.3
pinecone==3.0.0
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9

# PDF Processing
PyMuPDF==1.23.7

# GCP (Google Cloud Storage)
google-cloud-storage==2.14.0
google-auth==2.29.0

# Frontend UI
streamlit==1.32.2

# Other
requests==2.31.0
